// src/app/Model/AR Management/Workgroup/workgroup.ts
export class Workgroup {
  nworkqueuegroupid: string;
  sworkqueuegroupcode: string;
  clientcode: string;
  planname: string; // insurance
  claimno: string;
  dos: Date;
  remainingbalance: number; // insurance due
  claimage: number;
  lastworkeddate: Date;
  autolastworkeddate: Date;
  status: string; // automation status
  isassigned: boolean;
  payername: string;
  sausid: string;
  nclientid: string;
  sclientname: string;
}